Implemented a better way of passing options to the asynchronous ingest queue (ref #4865)
